add check style
