+++ src/main/java/junny/CommandTypes.java
@@ -1,4 +1,4 @@
-package Junny;
+package junny;
 
 public enum CommandTypes {
         TODO,
