+++ src/main/java/junny/command/EventCommand.java
@@ -1,11 +1,12 @@
-package Junny.Command;
-
-import Junny.Tasks.Event;
-import Junny.Storage;
-import Junny.Tasks.Task;
-import Junny.Ui;
+package junny.command;
 
 import java.util.ArrayList;
+
+import junny.Storage;
+import junny.Ui;
+import junny.tasks.Event;
+import junny.tasks.Task;
+
 
 public class EventCommand extends Command {
     private String description;
