+++ src/main/java/junny/command/TodoCommand.java
@@ -1,14 +1,14 @@
-package Junny.Command;
-
-import Junny.Storage;
-import Junny.Tasks.Task;
-import Junny.Tasks.Todo;
-import Junny.Ui;
+package junny.command;
 
 import java.util.ArrayList;
 
+import junny.Storage;
+import junny.Ui;
+import junny.tasks.Task;
+import junny.tasks.Todo;
+
 public class TodoCommand extends Command {
-    private  String description;
+    private String description;
 
     public TodoCommand(String description) {
         this.description = description;
