+++ src/main/java/junny/command/UnmarkCommand.java
@@ -1,10 +1,10 @@
-package Junny.Command;
-
-import Junny.Storage;
-import Junny.Tasks.Task;
-import Junny.Ui;
+package junny.command;
 
 import java.util.ArrayList;
+
+import junny.Storage;
+import junny.Ui;
+import junny.tasks.Task;
 
 public class UnmarkCommand extends Command {
     private final int index;
@@ -20,7 +20,9 @@ public class UnmarkCommand extends Command {
                 tasks.get(index - 1).markAsNotDone();
                 ui.markUndone(tasks.get(index - 1));
                 storage.saveAllTasks(tasks);
-            } else ui.printError("The task is not done yet, and you do not need to undo it!");
+            } else {
+                ui.printError("The task is not done yet, and you do not need to undo it!");
+            }
         } catch (NumberFormatException e) {
             // handle exception 3
             ui.printError("Please enter a valid number for unmark.");
@@ -29,5 +31,4 @@ public class UnmarkCommand extends Command {
             ui.printError("The task number you give does not exist. Please check again!");
         }
     }
-
 }
