+++ src/main/java/junny/tasks/Task.java
@@ -1,6 +1,6 @@
-package Junny.Tasks;
+package junny.tasks;
 
-import Junny.TaskTypes;
+import junny.TaskTypes;
 
 /**
  * Represents a task with a description, completion status and task type.
@@ -66,23 +66,29 @@ public abstract class Task {
         boolean isDone = getTaskType[1].equals("1");
         String task = getTaskType[2];
         switch(type) {
-            case "T":
-                Todo todo = new Todo(task);
-                if(isDone) todo.markAsDone();
-                return todo;
-            case "D":
-                Deadline ddl = new Deadline(task, getTaskType[3]);
-                if(isDone) ddl.markAsDone();
-                return ddl;
-            case "E":
-                String time = getTaskType[3];
-                // form is fromTo[0], to is fromTo[1]
-                String[] fromTo = time.split(" to ");
-                Event e = new Event(task, fromTo[0], fromTo[1]);
-                if(isDone) e.markAsDone();
-                return e;
-            default:
-                return null;
+        case "T":
+            Todo todo = new Todo(task);
+            if (isDone) {
+                todo.markAsDone();
+            }
+            return todo;
+        case "D":
+            Deadline ddl = new Deadline(task, getTaskType[3]);
+            if (isDone) {
+                ddl.markAsDone();
+            }
+            return ddl;
+        case "E":
+            String time = getTaskType[3];
+            // form is fromTo[0], to is fromTo[1]
+            String[] fromTo = time.split(" to ");
+            Event e = new Event(task, fromTo[0], fromTo[1]);
+            if (isDone) {
+                e.markAsDone();
+            }
+            return e;
+        default:
+            return null;
         }
     }
 }
