+++ src/test/java/Junny/SetupTest.java
@@ -1,4 +1,4 @@
-package Junny;
+package junny;
 
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertTrue;
