+++ text-ui-test/output.TXT
@@ -1,5 +1,5 @@
 ____________________________________________________________
- Hello! I'm Junny.Junny
+ Hello! I'm junny.junny
  What can I do for you?
 ____________________________________________________________
 ____________________________________________________________
